backend/requirements.txt
authlib==1.6.5
firebase_admin==7.1.0
Flask==3.1.2
Flask_Bcrypt==1.0.1
Flask_Login==0.6.3
flask_mail==0.10.0
flask_wtf==1.2.2
itsdangerous==2.2.0
python-dotenv==1.2.1
Werkzeug==3.1.5
WTForms==3.2.1
gunicorn
itsdangerous
cryptography
requests